modularized helper functions for visualize_program_progress() into helper_functions.py
